% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inherent_loss.R
\name{calc_lognorm_loss}
\alias{calc_lognorm_loss}
\title{Functions to compute inherent risk loss}
\usage{
calc_lognorm_loss(prob, upper, lower)
}
\arguments{
\item{prob}{probability of a loss.}

\item{upper}{upper bound of a 90% confidence interval of the impact in dollars.}

\item{lower}{lower bound of a 90% confidence interval of the impact in dollars.}
}
\value{
a vector of estimated losses in dollars.
}
\description{
Functions to compute inherent risk loss
}
\examples{
prob_loss <- c(.4, .15, .35, .1, .05, .12, .035)
impact_lower <- c(5e5, 2e6, 15e4, 2e5, 5e5, 5e5, 1e6)
impact_upper <- c(1e7, 5e6, 3e5, 5e5, 2e6, 1e7, 25e6)
calc_lognorm_loss(prob_loss, impact_upper, impact_lower)
}
